chore: Cleans up some things, flattens code structure, removes generic datamodel layer
