+++ src/data-model/storage-pointers/in-memory-json.js
