+++ test/wt-libs/data-model/wallet.spec.js
@@ -1,16 +1,16 @@
 import { assert } from 'chai';
 import sinon from 'sinon';
-import helpers from '../../../utils/helpers';
-import testedDataModel from '../../../utils/data-model-definition';
-import jsonWallet from '../../../utils/test-wallet';
-import Web3UriDataModel from '../../../../src/data-model/web3-uri';
-import Web3WTWallet from '../../../../src/data-model/web3-uri/wallet';
+import helpers from '../../utils/helpers';
+import testedDataModel from '../../utils/data-model-definition';
+import jsonWallet from '../../utils/test-wallet';
+import DataModel from '../../../src/data-model/';
+import Web3WTWallet from '../../../src/data-model/wallet';
 
-describe('WTLibs.data-model.web3-uri.wallet', () => {
+describe('WTLibs.data-model.wallet', () => {
   let dataModel;
   const correctPassword = 'test123';
   beforeEach(async function () {
-    dataModel = Web3UriDataModel.createInstance(testedDataModel.withDataSource().dataModelOptions);
+    dataModel = DataModel.createInstance(testedDataModel.withDataSource().dataModelOptions);
   });
 
   describe('unlock', () => {
